+++ src/test/java/wooteco/subway/admin/domain/LineTest.java
@@ -35,6 +35,6 @@ public class LineTest {
     void removeLineStation(Long stationId) {
         line.removeLineStationById(stationId);
 
-        assertThat(line.getStations()).hasSize(2);
+        assertThat(line.getStations().getStations()).hasSize(2);
     }
 }
