+++ src/test/java/wooteco/subway/admin/service/LineServiceTest.java
@@ -47,7 +47,7 @@ public class LineServiceTest {
         when(lineRepository.findById(line.getId())).thenReturn(Optional.of(line));
         lineService.addLineStation(line.getId(), request.toLineStation());
 
-        assertThat(line.getStations()).hasSize(4);
+        assertThat(line.getStations().getStations()).hasSize(4);
         assertThat(line.getStationIds().get(0)).isEqualTo(4L);
         assertThat(line.getStationIds().get(1)).isEqualTo(1L);
         assertThat(line.getStationIds().get(2)).isEqualTo(2L);
@@ -61,7 +61,7 @@ public class LineServiceTest {
         when(lineRepository.findById(line.getId())).thenReturn(Optional.of(line));
         lineService.addLineStation(line.getId(), request.toLineStation());
 
-        assertThat(line.getStations()).hasSize(4);
+        assertThat(line.getStations().getStations()).hasSize(4);
         assertThat(line.getStationIds().get(0)).isEqualTo(1L);
         assertThat(line.getStationIds().get(1)).isEqualTo(4L);
         assertThat(line.getStationIds().get(2)).isEqualTo(2L);
@@ -75,7 +75,7 @@ public class LineServiceTest {
         when(lineRepository.findById(line.getId())).thenReturn(Optional.of(line));
         lineService.addLineStation(line.getId(), request.toLineStation());
 
-        assertThat(line.getStations()).hasSize(4);
+        assertThat(line.getStations().getStations()).hasSize(4);
         assertThat(line.getStationIds().get(0)).isEqualTo(1L);
         assertThat(line.getStationIds().get(1)).isEqualTo(2L);
         assertThat(line.getStationIds().get(2)).isEqualTo(3L);
@@ -87,7 +87,7 @@ public class LineServiceTest {
         when(lineRepository.findById(line.getId())).thenReturn(Optional.of(line));
         lineService.removeLineStation(line.getId(), 1L);
 
-        assertThat(line.getStations()).hasSize(2);
+        assertThat(line.getStations().getStations()).hasSize(2);
         assertThat(line.getStationIds().get(0)).isEqualTo(2L);
         assertThat(line.getStationIds().get(1)).isEqualTo(3L);
     }
@@ -97,7 +97,7 @@ public class LineServiceTest {
         when(lineRepository.findById(line.getId())).thenReturn(Optional.of(line));
         lineService.removeLineStation(line.getId(), 2L);
 
-        assertThat(line.getStations()).hasSize(2);
+        assertThat(line.getStations().getStations()).hasSize(2);
         assertThat(line.getStationIds().get(0)).isEqualTo(1L);
         assertThat(line.getStationIds().get(1)).isEqualTo(3L);
     }
@@ -107,7 +107,7 @@ public class LineServiceTest {
         when(lineRepository.findById(line.getId())).thenReturn(Optional.of(line));
         lineService.removeLineStation(line.getId(), 3L);
 
-        assertThat(line.getStations()).hasSize(2);
+        assertThat(line.getStations().getStations()).hasSize(2);
         assertThat(line.getStationIds().get(0)).isEqualTo(1L);
         assertThat(line.getStationIds().get(1)).isEqualTo(2L);
     }
